The End

+++ src/main/java/com/tactfactory/demo/dtos/BookDto.java
@@ -3,7 +3,7 @@ package com.tactfactory.demo.dtos;
 public class BookDto {
 	
 	private String name;
-	private Long nbPage;
+	private int nbPage;
 	private Float price;
 	
 	public String getName() {
@@ -12,10 +12,10 @@ public class BookDto {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public Long getNbPage() {
+	public int getNbPage() {
 		return nbPage;
 	}
-	public void setNbPage(Long nbPage) {
+	public void setNbPage(int nbPage) {
 		this.nbPage = nbPage;
 	}
 	public Float getPrice() {
@@ -24,7 +24,5 @@ public class BookDto {
 	public void setPrice(Float price) {
 		this.price = price;
 	}
-	
-	
 
 }
